requirements.txt
attrs>=19.2.0
pytest~=5.3
pytest-cov~=2.8
pytest-mock~=1.13
pytruth~=1.1
hypothesis~=4.56